Remove DocumentTitleConverterFactory
